chore: misc adjustments and cleanup

+++ src/main/java/edu/jhuapl/trinity/data/audio/FourierTransform.java
@@ -193,7 +193,7 @@ public abstract class FourierTransform {
         }
 
         if (whichAverage == LINAVG) {
-            int avgWidth = (int) spectrum.length / averages.length;
+            int avgWidth = spectrum.length / averages.length;
             for (int i = 0; i < averages.length; i++) {
                 float avg = 0;
                 int j;
@@ -419,7 +419,7 @@ public abstract class FourierTransform {
     public float getAverageCenterFrequency(int i) {
         if (whichAverage == LINAVG) {
             // an average represents a certain number of bands in the spectrum
-            int avgWidth = (int) spectrum.length / averages.length;
+            int avgWidth = spectrum.length / averages.length;
             // the "center" bin of the average, this is fudgy.
             int centerBinIndex = i * avgWidth + avgWidth / 2;
             return indexToFreq(centerBinIndex);
